# loop-api/openapi.prod.json
{"openapi":"3.1.0","info":{"title":"Loop API","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/send_message":{"post":{"tags":["website-facade"],"summary":"Send Message","description":"Insert a human -> bot message in the given thread.\n\nBehaviour:\n- Validates thread and membership.\n- Inserts a row into messages with audience='inbox_to_bot'.\n- Returns the inserted row (plaintext content).","operationId":"send_message_api_send_message_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessagePayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/get_messages":{"get":{"tags":["website-facade"],"summary":"Get Messages","description":"Return a merged chronological stream for the viewer:\n\n- Human->Bot rows authored by this viewer (their own sent messages), AND\n- Bot->User rows targeted at this viewer (bot's DMs to them).\n\nThis allows the frontend to render:\n  - A or B (viewer’s human posts)      -> audience='inbox_to_bot'\n  - Bot→Viewer (personalised bot rows) -> audience='bot_to_user' AND recipient_profile_id=user_id\n\nResponse items include `audience` and `recipient_profile_id` so the UI can\nsplit panes for A, B, Bot→A, Bot→B if needed.","operationId":"get_messages_api_get_messages_get","parameters":[{"name":"thread_id","in":"query","required":true,"schema":{"type":"string","description":"Thread UUID","title":"Thread Id"},"description":"Thread UUID"},{"name":"user_id","in":"query","required":true,"schema":{"type":"string","description":"Profile UUID of the viewer whose DM stream we are showing","title":"User Id"},"description":"Profile UUID of the viewer whose DM stream we are showing"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetMessagesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/bot/process":{"post":{"tags":["bot"],"summary":"Process Queue","description":"Process human→bot messages and fan out bot→user DMs.\n\n- dry_run=True  → preview only (no DB writes; DO NOT mark processed)\n- dry_run=False → insert bot_to_user rows + mark source human with bot_processed_at","operationId":"process_queue_api_bot_process_post","parameters":[{"name":"thread_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Only process this thread","title":"Thread Id"},"description":"Only process this thread"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"dry_run","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Dry Run"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BotProcessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/whoami":{"get":{"tags":["diagnostics"],"summary":"Whoami","operationId":"whoami_api_whoami_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/bot/api/bot/process":{"post":{"tags":["bot"],"summary":"Process Queue","description":"Process human→bot messages and fan out bot→user DMs.\n\n- dry_run=True  → preview only (no DB writes; DO NOT mark processed)\n- dry_run=False → insert bot_to_user rows + mark source human with bot_processed_at","operationId":"process_queue_bot_api_bot_process_post","parameters":[{"name":"thread_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Only process this thread","title":"Thread Id"},"description":"Only process this thread"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"dry_run","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Dry Run"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BotProcessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/bot/api/bot/process":{"post":{"tags":["bot"],"summary":"Process Queue","description":"Process human→bot messages and fan out bot→user DMs.\n\n- dry_run=True  → preview only (no DB writes; DO NOT mark processed)\n- dry_run=False → insert bot_to_user rows + mark source human with bot_processed_at","operationId":"process_queue_api_bot_api_bot_process_post","parameters":[{"name":"thread_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Only process this thread","title":"Thread Id"},"description":"Only process this thread"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":10,"title":"Limit"}},{"name":"dry_run","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Dry Run"}},{"name":"X-User-Id","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-User-Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BotProcessResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health/dbinfo":{"get":{"summary":"Dbinfo","operationId":"dbinfo_health_dbinfo_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/feed/selftest":{"get":{"tags":["feed"],"summary":"Feed Selftest","operationId":"feed_selftest_api_feed_selftest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/feed":{"get":{"tags":["feed"],"summary":"Get Feed","operationId":"get_feed_api_feed_get","parameters":[{"name":"loop_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Loop Id"}},{"name":"for_profile_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"For Profile Id"}},{"name":"since","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Since"}},{"name":"preview","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Preview"}},{"name":"last_seen_hours","in":"query","required":false,"schema":{"type":"integer","maximum":720,"minimum":1,"default":48,"title":"Last Seen Hours"}},{"name":"max_messages","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Max Messages"}},{"name":"include_self","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Include Self"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedDigest"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BotProcessItem":{"properties":{"human_message_id":{"type":"string","title":"Human Message Id","description":"source inbox_to_bot message id"},"thread_id":{"type":"string","title":"Thread Id"},"recipients":{"items":{"type":"string"},"type":"array","title":"Recipients","default":[]},"bot_rows":{"items":{"type":"string"},"type":"array","title":"Bot Rows","default":[]},"previews":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array","title":"Previews","default":[]},"skipped_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skipped Reason"}},"type":"object","required":["human_message_id","thread_id"],"title":"BotProcessItem"},"BotProcessResponse":{"properties":{"ok":{"type":"boolean","title":"Ok","default":true},"reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Reason"},"stats":{"$ref":"#/components/schemas/BotProcessStats"},"items":{"items":{"$ref":"#/components/schemas/BotProcessItem"},"type":"array","title":"Items","default":[]}},"type":"object","required":["stats"],"title":"BotProcessResponse"},"BotProcessStats":{"properties":{"scanned":{"type":"integer","title":"Scanned","default":0},"processed":{"type":"integer","title":"Processed","default":0},"inserted":{"type":"integer","title":"Inserted","default":0},"skipped":{"type":"integer","title":"Skipped","default":0},"dry_run":{"type":"boolean","title":"Dry Run","default":true}},"type":"object","title":"BotProcessStats"},"FeedDigest":{"properties":{"loop_id":{"type":"string","format":"uuid","title":"Loop Id"},"for_profile_id":{"type":"string","format":"uuid","title":"For Profile Id"},"items_count":{"type":"integer","title":"Items Count"},"window_start":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Window Start"},"window_end":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Window End"},"digest_text":{"type":"string","title":"Digest Text"},"last_seen_at_prev":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Seen At Prev"},"last_seen_at_new":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Seen At New"},"engine":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Engine"}},"type":"object","required":["loop_id","for_profile_id","items_count","digest_text"],"title":"FeedDigest"},"GetMessagesResponse":{"properties":{"ok":{"type":"boolean","title":"Ok","default":true},"items":{"items":{"$ref":"#/components/schemas/MessageOut"},"type":"array","title":"Items","default":[]}},"type":"object","title":"GetMessagesResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageOut":{"properties":{"id":{"type":"string","title":"Id"},"thread_id":{"type":"string","title":"Thread Id"},"created_at":{"type":"string","title":"Created At"},"created_by":{"type":"string","title":"Created By"},"author_member_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author Member Id"},"audience":{"type":"string","title":"Audience"},"recipient_profile_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Recipient Profile Id"},"content":{"type":"string","title":"Content"}},"type":"object","required":["id","thread_id","created_at","created_by","audience","content"],"title":"MessageOut"},"SendMessagePayload":{"properties":{"thread_id":{"type":"string","title":"Thread Id","description":"UUID of the thread"},"user_id":{"type":"string","title":"User Id","description":"Profile UUID of the human sender"},"content":{"type":"string","maxLength":4000,"minLength":1,"title":"Content"}},"type":"object","required":["thread_id","user_id","content"],"title":"SendMessagePayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}